+++ homeassistant/components/bthome/const.py
@@ -1,3 +1,3 @@
 """Constants for the BThome Bluetooth integration."""
 
-DOMAIN = "bthome_ble"
+DOMAIN = "bthome"
