+++ homeassistant/components/bthome/strings.json
