+++ homeassistant/components/bthome/translations/en.json
