+++ requirements_all.txt
@@ -457,7 +457,7 @@ bsblan==0.5.0
 # homeassistant.components.bluetooth_tracker
 bt_proximity==0.2.1
 
-# homeassistant.components.bthome_ble
+# homeassistant.components.bthome
 bthome-ble==0.3.4
 
 # homeassistant.components.bt_home_hub_5
