+++ tests/components/bthome/conftest.py
