+++ tests/components/bthome/test_sensor.py
@@ -3,7 +3,7 @@
 from unittest.mock import patch
 
 from homeassistant.components.bluetooth import BluetoothChange
-from homeassistant.components.bthome_ble.const import DOMAIN
+from homeassistant.components.bthome.const import DOMAIN
 from homeassistant.components.sensor import ATTR_STATE_CLASS
 from homeassistant.const import ATTR_FRIENDLY_NAME, ATTR_UNIT_OF_MEASUREMENT
